requirements.txt
apache-airflow==2.10.5
confluent-kafka==2.8.0
python-dotenv== 1.1.0
pymssql==2.3.2
pyarrow==20.0.0
pandas==2.1.3
pymongo
discordwebhook==1.0.3